+++ sickbeard/helpers.py
@@ -1432,6 +1432,8 @@ def get_size(start_path='.'):
 def remove_article(text=''):
     return re.sub(r'(?i)^(?:(?:A(?!\s+to)n?)|The)\s(\w)', r'\1', text)
 
+def build_dict(seq, key):
+    return dict((d[key], dict(d, index=index)) for (index, d) in enumerate(seq))
 
 def client_host(server_host):
     '''Extracted from cherrypy libs